requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
groq>=1.0.0